Fix SSL issues when installing the agent on older operating systems
Some older operating systems' certificates can't be updated via their package
managers, so the `step.sm` domain was seen as invalid when attempting to
download the step client. This patch changes the download URL domain to
the `github.com` one, tries to download without checking the
certificates, in case it fails the first time.